Google Scripts that handled credit card tips
These scripts were used to organzie the input of tips after closing. Made a new sheet for everyweek and autoset the new dates. Definitely could have several more improvements but they work as they are and I wasn't in charge of handing tips anymore so this is what I gave the receptionist to use.